+++ syncserver-db-common/Cargo.toml
@@ -4,21 +4,15 @@ version = "0.12.5"
 edition = "2021"
 
 [dependencies]
-async-trait = "0.1.40"
+actix-web = "3"
 backtrace = "0.3.61"
-chrono = "0.4"
 # Pin to 0.5 for now, to keep it under tokio 0.2 (issue977).
 # Fix for #803 (deadpool#92) points to our fork for now
 #deadpool = "0.5"  # pin to 0.5
 deadpool = { git = "https://github.com/mozilla-services/deadpool", branch = "deadpool-v0.5.2-issue92" }
 diesel = { version = "1.4", features = ["mysql", "r2d2"] }
 diesel_migrations = { version = "1.4.0", features = ["mysql"] }
-http = "0.2.6"
 futures = { version = "0.3", features = ["compat"] }
-lazy_static = "1.4.0"
-serde = "1.0"
-serde_json = { version = "1.0", features = ["arbitrary_precision"] }
+http = "0.2.6"
 syncserver-common = { path = "../syncserver-common" }
 thiserror = "1.0.26"
-# pinning to 0.2.4 due to high number of dependencies (actix, bb8, deadpool, etc.)
-tokio = { version = "0.2.4", features = ["macros", "sync", "blocking"] }
