+++ syncserver/Cargo.toml
@@ -60,6 +60,7 @@ slog-stdlog = "4.1"
 slog-term = "2.6"
 syncserver-settings = { path = "../syncserver-settings" }
 syncserver-db-common = { path = "../syncserver-db-common" }
+syncstorage-db-common = { path = "../syncstorage-db-common" }
 syncserver-common = { path = "../syncserver-common" }
 syncstorage-mysql = { path = "../syncstorage-mysql", optional = true }
 syncstorage-settings = { path = "../syncstorage-settings" }
