+++ syncserver/src/db/tests/support.rs
@@ -1,10 +1,10 @@
 use std::str::FromStr;
 
 use syncserver_common::Metrics;
-use syncserver_db_common::{
+use syncserver_settings::Settings as SyncserverSettings;
+use syncstorage_db_common::{
     params, util::SyncTimestamp, DbPool as DbPoolTrait, Sorting, UserIdentifier,
 };
-use syncserver_settings::Settings as SyncserverSettings;
 use syncstorage_settings::Settings as SyncstorageSettings;
 
 use crate::db::{BoxDb, DbPool};
