+++ syncserver/src/error.rs
@@ -21,8 +21,8 @@ use serde::{
     Serialize,
 };
 
-use syncserver_common::{from_error, impl_fmt_display, ReportableError};
-use syncserver_db_common::error::DbErrorIntrospect;
+use syncserver_common::{from_error, impl_fmt_display, MetricError, ReportableError};
+use syncstorage_db_common::error::DbErrorIntrospect;
 
 #[cfg(feature = "mysql")]
 use syncstorage_mysql::error::DbError;
