+++ syncserver/src/server/test.rs
@@ -16,12 +16,12 @@ use serde::de::DeserializeOwned;
 use serde_json::json;
 use sha2::Sha256;
 use syncserver_common::{self, X_LAST_MODIFIED};
-use syncserver_db_common::{
+use syncserver_settings::{Secrets, Settings};
+use syncstorage_db_common::{
     params,
     results::{DeleteBso, GetBso, PostBsos, PutBso},
     util::SyncTimestamp,
 };
-use syncserver_settings::{Secrets, Settings};
 use syncstorage_settings::ServerLimits;
 
 use super::*;
