+++ syncserver/src/web/middleware/weave.rs
@@ -10,7 +10,7 @@ use actix_web::{
 
 use futures::future::{self, LocalBoxFuture};
 use syncserver_common::{X_LAST_MODIFIED, X_WEAVE_TIMESTAMP};
-use syncserver_db_common::util::SyncTimestamp;
+use syncstorage_db_common::util::SyncTimestamp;
 
 use crate::error::{ApiError, ApiErrorKind};
 use crate::web::DOCKER_FLOW_ENDPOINTS;
