+++ syncstorage-db-common/src/results.rs
