+++ tokenserver-common/Cargo.toml
@@ -9,5 +9,4 @@ backtrace = "0.3.61"
 serde = "1.0"
 serde_json = { version = "1.0", features = ["arbitrary_precision"] }
 syncserver-common = { path = "../syncserver-common" }
-syncserver-db-common = { path = "../syncserver-db-common" }
 thiserror = "1.0.26"
